@@ -1,74 +1,76 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ExtendedLog.cs" company="Marcin Smółka zNET Computer Solutions">
-//   Copyright (c) Marcin Smółka zNET Computer Solutions. All rights reserved.
+// <copyright file="ExtendedLog.cs" company="Marcin Smółka">
+//   Copyright (c) Marcin Smółka. All rights reserved.
 // </copyright>
 // <summary>
 //   The extended log.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace ZNetCS.AspNetCore.Logging.EntityFrameworkCoreTest
+namespace ZNetCS.AspNetCore.Logging.EntityFrameworkCoreTest;
+
+#region Usings
+
+using System.Diagnostics.CodeAnalysis;
+
+using Microsoft.AspNetCore.Http;
+
+using ZNetCS.AspNetCore.Logging.EntityFrameworkCore;
+
+#endregion
+
+/// <summary>
+/// The extended log.
+/// </summary>
+[SuppressMessage("ReSharper", "ClassNeverInstantiated.Global", Justification = "Test")]
+public class ExtendedLog : Log
 {
-    #region Usings
+    #region Constructors and Destructors
 
-    using Microsoft.AspNetCore.Http;
+    /// <summary>
+    /// Initializes a new instance of the <see cref="ExtendedLog"/> class.
+    /// </summary>
+    /// <param name="accessor">
+    /// The accessor.
+    /// </param>
+    public ExtendedLog(IHttpContextAccessor accessor)
+    {
+        this.Browser = "Test Browser";
+        this.Host = "localhost";
+        this.User = "Test User";
+        this.Path = accessor.HttpContext?.Request.Path;
+    }
 
-    using ZNetCS.AspNetCore.Logging.EntityFrameworkCore;
+    /// <summary>
+    /// Initializes a new instance of the <see cref="ExtendedLog"/> class.
+    /// </summary>
+    protected ExtendedLog()
+    {
+    }
 
     #endregion
 
+    #region Public Properties
+
     /// <summary>
-    /// The extended log.
+    /// Gets or sets the browser.
     /// </summary>
-    public class ExtendedLog : Log
-    {
-        #region Constructors and Destructors
+    public string? Browser { get; set; }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ExtendedLog"/> class.
-        /// </summary>
-        /// <param name="accessor">
-        /// The accessor.
-        /// </param>
-        public ExtendedLog(IHttpContextAccessor accessor)
-        {
-            this.Browser = "Test Browser";
-            this.Host = "localhost";
-            this.User = "Test User";
-            this.Path = accessor.HttpContext.Request.Path;
-        }
+    /// <summary>
+    /// Gets or sets the host.
+    /// </summary>
+    public string? Host { get; set; }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ExtendedLog"/> class.
-        /// </summary>
-        protected ExtendedLog()
-        {
-        }
+    /// <summary>
+    /// Gets or sets the path.
+    /// </summary>
+    public string? Path { get; set; }
 
-        #endregion
+    /// <summary>
+    /// Gets or sets the user.
+    /// </summary>
+    public string? User { get; set; }
 
-        #region Public Properties
-
-        /// <summary>
-        /// Gets or sets the browser.
-        /// </summary>
-        public string Browser { get; set; }
-
-        /// <summary>
-        /// Gets or sets the host.
-        /// </summary>
-        public string Host { get; set; }
-
-        /// <summary>
-        /// Gets or sets the path.
-        /// </summary>
-        public string Path { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user.
-        /// </summary>
-        public string User { get; set; }
-
-        #endregion
-    }
+    #endregion
 }
