@@ -1,74 +1,62 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="TestLoggerBuilderSimpleNoFilter.cs" company="Marcin Smółka zNET Computer Solutions">
-//   Copyright (c) Marcin Smółka zNET Computer Solutions. All rights reserved.
+// <copyright file="TestLoggerBuilderSimpleNoFilter.cs" company="Marcin Smółka">
+//   Copyright (c) Marcin Smółka. All rights reserved.
 // </copyright>
 // <summary>
 //   The test logger simple.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace ZNetCS.AspNetCore.Logging.EntityFrameworkCoreTest
+namespace ZNetCS.AspNetCore.Logging.EntityFrameworkCoreTest;
+
+#region Usings
+
+using System.Linq;
+using System.Net.Http;
+using System.Threading.Tasks;
+
+using Microsoft.AspNetCore.TestHost;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+#endregion
+
+/// <summary>
+/// The test logger simple.
+/// </summary>
+[TestClass]
+public class TestLoggerBuilderSimpleNoFilter
 {
-    #region Usings
-
-    using System.Linq;
-    using System.Net.Http;
-    using System.Threading.Tasks;
-
-    using Microsoft.AspNetCore.TestHost;
-    using Microsoft.EntityFrameworkCore;
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-    #endregion
+    #region Public Methods
 
     /// <summary>
-    /// The test logger simple.
+    /// The write log.
     /// </summary>
-    [TestClass]
-    public class TestLoggerBuilderSimpleNoFilter : TestBuilderBase
+    [TestMethod]
+    public async Task WriteBuilderSimpleNoFilterLog()
     {
-        #region Constructors and Destructors
+        using var factory = new TestWebApplicationFactory(TestVersion.BuilderSimpleNoFilter);
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TestLoggerBuilderSimpleNoFilter"/> class.
-        /// </summary>
-        public TestLoggerBuilderSimpleNoFilter() : base(4)
-        {
-        }
+        var options = new DbContextOptionsBuilder<ContextSimple>()
+            .UseInMemoryDatabase("SimpleLogNoFilterDatabase", StartupBuilderSimpleNoFilter.MemoryRoot)
+            .Options;
 
-        #endregion
+        // Act
+        RequestBuilder request = factory.Server.CreateRequest("/");
+        HttpResponseMessage response = await request.SendAsync("PUT");
 
-        #region Public Methods
+        // Assert
+        response.EnsureSuccessStatusCode();
 
-        /// <summary>
-        /// The write log.
-        /// </summary>
-        [TestMethod]
-        public async Task WriteBuilderSimpleNoFilterLog()
-        {
-            var options = new DbContextOptionsBuilder<ContextSimple>()
-                .UseInMemoryDatabase("SimpleLogNoFilterDatabase", StartupBuilderSimpleNoFilter.MemoryRoot)
-                .Options;
+        // Use a separate instance of the context to verify correct data was saved to database
+        await using var context = new ContextSimple(options);
+        var logs = context.Logs.ToList();
 
-            // Act
-            RequestBuilder request = this.Server.CreateRequest("/");
-            HttpResponseMessage response = await request.SendAsync("PUT");
-
-            // Assert
-            response.EnsureSuccessStatusCode();
-
-            // Use a separate instance of the context to verify correct data was saved to database
-            using (var context = new ContextSimple(options))
-            {
-                var logs = context.Logs.ToList();
-
-                Assert.AreEqual("Handling request.", logs[1].Message);
-                Assert.AreEqual(1, logs[1].EventId);
-                Assert.AreEqual("Finished handling request.", logs[2].Message);
-                Assert.AreEqual(2, logs[2].EventId);
-            }
-        }
-
-        #endregion
+        Assert.AreEqual("Handling request", logs[0].Message);
+        Assert.AreEqual(1, logs[0].EventId);
+        Assert.AreEqual("Finished handling request", logs[1].Message);
+        Assert.AreEqual(2, logs[1].EventId);
     }
+
+    #endregion
 }
