@@ -1,6 +1,6 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="AssemblyInfo.cs" company="Marcin Smółka zNET Computer Solutions">
-//   Copyright (c) Marcin Smółka zNET Computer Solutions. All rights reserved.
+// <copyright file="AssemblyInfo.cs" company="Marcin Smółka">
+//   Copyright (c) Marcin Smółka. All rights reserved.
 // </copyright>
 // <summary>
 //   AssemblyInfo.cs
@@ -9,16 +9,8 @@
 
 #region Usings
 
-using System.Reflection;
 using System.Runtime.InteropServices;
 
 #endregion
 
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Marcin Smółka zNET Computer Solutions")]
-[assembly: AssemblyTitle("ZNetCS.AspNetCore.Logging.EntityFrameworkCore")]
-[assembly: AssemblyProduct("ZNetCS.AspNetCore.Logging.EntityFrameworkCore")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCopyright("Copyright © Marcin Smółka zNET Computer Solutions 2016 - 2017")]
-[assembly: ComVisible(false)]
 [assembly: Guid("f56f98eb-7a88-4f81-ac08-e06f8265b142")]
